adds test files and codes
